requirements.txt
Flask==1.1.1
Flask-JWT-Extended==3.22.0
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.13.0
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.8.5
Flask-Migrate==2.5.2
Flask-Mail==0.9.1
Flask-Cors==3.0.8
sentry-sdk==0.10.2
blinker==1.4
uwsgi==2.0.18
